Split AdmingBaseTestCase outside of test_admin_page.py.
This shuts up the linter complaining about the module having too many lines.
